uiunt SetZeroAtPointerX

+++ BackendX64Function_.cpp
@@ -1460,7 +1460,7 @@ bool BackendX64::emitFunctionBody(const BuildParameters& buildParameters, Module
             break;
         case ByteCodeOp::SetZeroAtPointerX:
             SWAG_ASSERT(sizeParamsStack >= 3 * sizeof(Register));
-            BackendX64Inst::emit_Load64_Immediate(pp, ip->b.u32, RAX);
+            BackendX64Inst::emit_Load64_Immediate(pp, ip->b.u64, RAX);
             BackendX64Inst::emit_Store64_Indirect(pp, 16, RAX, RSP);
             BackendX64Inst::emit_Clear64(pp, RAX);
             BackendX64Inst::emit_Store64_Indirect(pp, 8, RAX, RSP);
